+++ src/components/PlayerProgress.tsx
@@ -5,7 +5,7 @@ import '../styles/PlayerProgress.css';
 
 const PlayerProgress: React.FC<PlayerProgressProps> = ({active, resolution}) => {
   return <div className="player_progress">
-    <div className={`player_resoultion ${active ? "selected" : ""}`}>{active? AppConfig.ResolutionsMap[resolution] : "Resolution"}
+    <div className={`player_resoultion ${active ? "selected" : ""}`}>{active? AppConfig.ResolutionsMap[resolution] : "Quality"}
     </div>
   </div>;
 };