+++ src/types/VideoPlayer.d.ts
@@ -1,7 +1,8 @@
 // Define types for the props
 interface VideoPlayerProps {
   mpdUrl: string; // URL to the MPD file
-  drmLicenseUrl: string
+  drmLicenseUrl: string,
+  displayProgress : boolean,
 } 
 
 export default VideoPlayerProps;